+++ circuit_maintenance_parser/parsers/lumen.py
@@ -15,14 +15,8 @@ from circuit_maintenance_parser.parser import Html, Impact, CircuitImpact, Statu
 logger = logging.getLogger(__name__)
 
 
-class ParserLumen(Html):
+class HtmlParserLumen1(Html):
     """Notifications Parser for Lumen notifications."""
-
-    provider_type: str = "lumen"
-
-    # Default values for Lumen notifications
-    _default_provider = "lumen"
-    _default_organizer = "smc@lumen.com"
 
     def parse_html(self, soup, data_base):
         """Execute parsing."""